+++ example/src/main/resources/META-INF/persistence.xml
@@ -4,7 +4,7 @@
              version="2.0">
 
     <persistence-unit name="example">
-        <provider>org.hibernate.reactive.jpa.impl.ReactivePersistenceProvider</provider>
+        <provider>org.hibernate.reactive.provider.ReactivePersistenceProvider</provider>
 
         <class>org.hibernate.example.reactive.Author</class>
         <class>org.hibernate.example.reactive.Book</class>
