+++ hibernate-reactive-core/src/main/resources/META-INF/services/javax.persistence.spi.PersistenceProvider
@@ -1,1 +1,1 @@
-org.hibernate.reactive.jpa.impl.ReactivePersistenceProvider
+org.hibernate.reactive.provider.ReactivePersistenceProvider
